Numoerous safe conversions implemented. Added additional Utility::convertSizeToDWORD for windows builds.
(cherry picked from commit d6af025a46b06637f3619ff12fe19ad86e7579cf)

+++ src/csync/vio/csync_vio_local_win.cpp
@@ -57,7 +57,7 @@ csync_vio_handle_t *csync_vio_local_opendir(const char *name) {
   handle = (dhandle_t*)c_malloc(sizeof(dhandle_t));
 
   // the file wildcard has to be attached
-  int len_name = strlen(name);
+  size_t len_name = strlen(name);
   if( len_name ) {
       char *h = NULL;
